migrated from `CategorySwap` to `TradeCategory` in models, repositories, services, and related APIs, and updated all references and ID types accordingly

+++ Modulith/src/main/java/com/tskhra/modulith/trade_module/controllers/ItemSearchController.java
@@ -30,7 +30,7 @@ public class ItemSearchController {
     @GetMapping
     public ResponseEntity<Page<ItemSummaryDto>> searchItems(
             @RequestParam(required = false) String query,
-            @RequestParam(required = false) Long categoryId,
+            @RequestParam(required = false) Integer categoryId,
             @RequestParam(required = false) Long cityId,
             @RequestParam(required = false) ItemCondition condition,
             @RequestParam(required = false) TradeRange tradeRange,
